Move arithmetic data supplier to core
